+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Latest first.
   - Returns true if any town in the supplied nation contains a cell under attack.
 - Add `OutlawListener` and listen for the `OutlawTeleportEvent`
   - Outlaws cannot teleport away during an attack while in enemy territory or enemy nation zone during a war.
+  - Add `error.outlaw.cannot-teleport-here` translation key, sent to the outlaw.
 
 ### Misc. Changes
 - Bump Towny API dependency to 0.97.0.17; set as Min_Towny_Version